# internal/docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "XYZ Books API",
        "title": "XYZ Books API",
        "contact": {
            "name": "Emilio Gozo",
            "email": "emiliogozo@proton.me"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/books": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "List books",
                "parameters": [
                    {
                        "type": "string",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "max_publication_year",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "min_publication_year",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 30,
                        "minimum": 1,
                        "type": "integer",
                        "description": "limit",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "publisher",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedBooks"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Create book",
                "parameters": [
                    {
                        "description": "Create book parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateBookParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                }
            }
        },
        "/books/{isbn}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Get book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISBN-13",
                        "name": "isbn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Update book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISBN-13",
                        "name": "isbn",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update book parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateBookParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Delete book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISBN-13",
                        "name": "isbn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Book": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "edition": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "isbn10": {
                    "type": "string"
                },
                "isbn13": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "publication_year": {
                    "type": "integer"
                },
                "publisher": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "CreateBookParams": {
            "type": "object",
            "required": [
                "authors",
                "publisher"
            ],
            "properties": {
                "authors": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "book": {
                    "type": "object",
                    "required": [
                        "price",
                        "publication_year",
                        "title"
                    ],
                    "properties": {
                        "edition": {
                            "type": "string"
                        },
                        "image_url": {
                            "type": "string"
                        },
                        "isbn10": {
                            "type": "string"
                        },
                        "isbn13": {
                            "type": "string"
                        },
                        "price": {
                            "type": "number"
                        },
                        "publication_year": {
                            "type": "integer",
                            "minimum": 1000
                        },
                        "title": {
                            "type": "string"
                        }
                    }
                },
                "publisher": {
                    "type": "string"
                }
            }
        },
...
        "PaginatedBooks": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Book"
                    }
                },
                "next_page": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "prev_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "UpdateBookParams": {
            "type": "object",
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "isbn10": {
                    "type": "string"
                },
                "isbn13": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "publication_year": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
    }
}